rattler::install::link

Function copy_and_replace_textual_placeholder

pub fn copy_and_replace_textual_placeholder(
    source_bytes: &[u8],
    destination: impl Write,
    prefix_placeholder: &str,
    target_prefix: &str,
    target_platform: &Platform,
) -> Result<(), Error>

Given the contents of a file copy it to the destination and in the process replace the prefix_placeholder text with the target_prefix text.

This is a text based version where the complete string is replaced. This works fine for text files but will not work correctly for binary files where the length of the string is often important. See copy_and_replace_cstring_placeholder when you are dealing with binary content.