rattler::install::link

Function link_file

pub fn link_file(
    path_json_entry: &PathsEntry,
    destination_relative_path: PathBuf,
    package_dir: &Path,
    target_dir: &Prefix,
    target_prefix: &str,
    allow_symbolic_links: bool,
    allow_hard_links: bool,
    allow_ref_links: bool,
    target_platform: Platform,
    apple_codesign_behavior: AppleCodeSignBehavior,
    modification_time: FileTime,
    external_symlink_policy: ExternalSymlinkPolicy,
) -> Result<Option<LinkedFile>, LinkFileError>

Installs a single file from a package_dir to the the target_dir. Replaces any prefix_placeholder in the file with the prefix.

relative_path is the path of the file in the package_dir (and the target_dir).

Note that usually the target_prefix is equal to target_dir but it might differ. See crate::install::InstallOptions::target_prefix for more information.

The modification_time is a timestamp we set on all files we modify. We want a value we control here to make the generated filesystem tree more reproducible. modification_time should be greater than any of the modification times of any of the files that were packaged up (ignoring any data conda stores).