rattler::install

Struct Transaction

pub struct Transaction<Old, New> {
    pub operations: Vec<TransactionOperation<Old, New>>,
    pub python_info: Option<PythonInfo>,
    pub current_python_info: Option<PythonInfo>,
    pub platform: Platform,
    pub unchanged: Vec<Old>,
}

Describes the operations to perform to bring an environment from one state into another.

Fields
operations: Vec<TransactionOperation<Old, New>>

A list of operations to update an environment

python_info: Option<PythonInfo>

The python version of the target state, or None if python doesn’t exist in the environment.

current_python_info: Option<PythonInfo>

The python version of the current state, or None if python didn’t exist in the previous environment.

platform: Platform

The target platform of the transaction

unchanged: Vec<Old>

The records that are not touched by the transaction.

Implementations

impl<Old: Clone, New: Clone> Transaction<&Old, &New>

pub fn to_owned(self) -> Transaction<Old, New>

Own records.

impl<Old, New> Transaction<Old, New>

pub fn removed_packages(&self) -> impl Iterator<Item = &Old> + '_

Return an iterator over the prefix records of all packages that are going to be removed.

pub fn unchanged_packages(&self) -> &[Old]

Return an iterator over the records that are not touched by the transaction

pub fn packages_to_uninstall(&self) -> usize

Returns the number of records to install.

impl<Old, New> Transaction<Old, New>

pub fn installed_packages(&self) -> impl Iterator<Item = &New> + '_

Return an iterator over the prefix records of all packages that are going to be installed.

pub fn packages_to_install(&self) -> usize

Returns the number of records to install.

impl<Old, New> Transaction<Old, New>
where Old: ContentComparable, New: ContentComparable,

pub fn from_current_and_desired<CurIter: IntoIterator<Item = Old>, NewIter: IntoIterator<Item = New>>( current: CurIter, desired: NewIter, reinstall: Option<&HashSet<PackageName>>, ignored: Option<&HashSet<PackageName>>, platform: Platform, ) -> Result<Self, TransactionError>

Constructs a Transaction by taking the current situation and diffing that against the desired situation. You can specify a set of package names that should be reinstalled even if their content has not changed. You can also specify a set of package names that should be ignored (left untouched).

impl<New> Transaction<InstallationResultRecord, New>

pub fn into_prefix_record( self, prefix: impl AsRef<Path>, ) -> Result<Transaction<PrefixRecord, New>, Error>

Convert transaction to contain PrefixRecord instead of InstallationResultRecord.

impl<New> Transaction<PrefixRecord, New>

pub fn into_installation_result_record( self, ) -> Transaction<InstallationResultRecord, New>

Convert transaction to contain InstallationResultRecord instead of PrefixRecord.