rattler

Module package_cache

This module provides functionality to cache extracted Conda packages. See PackageCache.

Structs

BucketKey

A key that defines the actual location of the package in the cache.

CacheGlobalLock

A global lock for the entire package cache.

CacheKey

Provides a unique identifier for packages in the cache. TODO: This could not be unique over multiple subdir. How to handle?

CacheMetadata

A validated cache entry with its associated metadata.

PackageCache

A PackageCache manages a cache of extracted Conda packages on disk.

PackageCacheLayer

Enums

PackageCacheError

Errors specific to the PackageCache interface

PackageCacheLayerError

Errors specific to individual layers in the PackageCache

Traits

CacheReporter

A trait that can be implemented to report progress of the download and validation process.