rattler::validation

Function validate_package_directory_from_paths

pub fn validate_package_directory_from_paths(
    package_dir: &Path,
    paths: &PathsJson,
    mode: ValidationMode,
) -> Result<(), (PathBuf, PackageEntryValidationError)>

Determine whether the files in the specified directory match wat is expected according to the passed in PathsJson.